[feat]RGBColorInProgressCircle 파일 삭제
