+++ ToGather/ToGather/Global/Extension/Color+Extension.swift
@@ -57,3 +57,18 @@ extension Color {
     static let friendColors: [Color] = [.friendRed01, .friendPurple01, .friendGreen01]
 }
 
+extension Color {
+    var components: (red: CGFloat, green: CGFloat, blue: CGFloat) {
+        
+        var r: CGFloat = 0
+        var g: CGFloat = 0
+        var b: CGFloat = 0
+        var o: CGFloat = 0
+        
+        guard UIColor(self).getRed(&r, green: &g, blue: &b, alpha: &o) else {
+            return (0, 0, 0)
+        }
+        
+        return (r, g, b)
+    }
+}
