--- ToGather/ToGather/Screens/Core/Main/UIComponent/ProgressCircle/RGBColorInProgressCircle.swift
@@ -1,26 +1,0 @@
-//
-//  RGBColorInProgressCircle.swift
-//  ToGather
-//
-//  Created by hyo on 2022/06/12.
-//
-
-import SwiftUI
-
-import UIKit
-
-extension Color {
-    var components: (red: CGFloat, green: CGFloat, blue: CGFloat) {
-        
-        var r: CGFloat = 0
-        var g: CGFloat = 0
-        var b: CGFloat = 0
-        var o: CGFloat = 0
-        
-        guard UIColor(self).getRed(&r, green: &g, blue: &b, alpha: &o) else {
-            return (0, 0, 0)
-        }
-        
-        return (r, g, b)
-    }
-}
